% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_base_dcf.R
\name{add}
\alias{add}
\title{Add dependencies to a DESCRIPTION file}
\usage{
add(replacements = c("utils", "stats", "grDevices"), field = "Imports",
  file = "DESCRIPTION", name = "myAnalysis", write = FALSE)
}
\arguments{
\item{replacements}{A character string of packages to add/replace}

\item{field}{Which field should be modified (defaults to "Imports")}

\item{file}{The DESCRIPTION file.}

\item{name}{The name of the project (in the case that the DESCRIPTION does not exist.)}

\item{write}{a logical. If `TRUE`, DESCRIPTION file will be overwritten, if `FALSE`, the file will be printed to screen.}
}
\value{
if `write = TRUE`, NULL, if `write = FALSE`, the DESCRIPTION file will be printed to screen.
}
\description{
Add dependencies to a DESCRIPTION file
}
\examples{
# a superficial example with this package's DESCRIPTION file
LIB <- .libPaths()[1]
pkg <- file.path(LIB, "diaper", "DESCRIPTION")
# note the Dependency version for R
write.dcf(read.dcf(pkg))
# now we change it to be 3.0
add(replacements = "R (>= 3.0.0)", field = "Depends",
    file = pkg, write = FALSE)
}
