% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_base_dcf.R
\name{remove}
\alias{remove}
\title{Remove dependencies from a DESCRIPTION file}
\usage{
remove(to_remove = c("utlis"), field = "Imports", file = "DESCRIPTION",
  write = FALSE)
}
\arguments{
\item{to_remove}{a list of dependencies to remove}

\item{field}{Which field should be modified (defaults to "Imports")}

\item{file}{The DESCRIPTION file.}

\item{write}{a logical. If `TRUE`, DESCRIPTION file will be overwritten, if `FALSE`, the file will be printed to screen.}
}
\value{
if `write = TRUE`, the specified file will be modified. If `write =
  FALSE`, the proposed changes.
}
\description{
This function will remove any entries that don't belong in the description
file, regardless of version number. This can be used to remove misspelled or
unused dependencies.
}
\examples{

tmp <- tempdir()
add(file = file.path(tmp, "DESCRIPTION"), write = TRUE)
write.dcf(read.dcf(file.path(tmp, "DESCRIPTION")))
remove("utils", file = file.path(tmp, "DESCRIPTION"))

}
