% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_base_dcf.R
\name{replace}
\alias{replace}
\title{Replace a dependency in a DESCRIPTION file}
\usage{
replace(from = "utlis", to = "utils", field = "Imports",
  file = "DESCRIPTION", write = FALSE)
}
\arguments{
\item{from}{a dependency to remove (sans parenthetical version)}

\item{to}{the depenency to replace (parenthetical version optional)}

\item{field}{Which field should be modified (defaults to "Imports")}

\item{file}{The DESCRIPTION file.}

\item{write}{a logical. If `TRUE`, DESCRIPTION file will be overwritten, if `FALSE`, the file will be printed to screen.}
}
\value{
if `write = TRUE`, the specified file will be modified. If `write =
  FALSE`, the proposed changes.
}
\description{
This will replace a dependency in a DESCRIPTION file, ignoring the parens
}
\examples{

tmp <- tempdir()
add(c("utils", "stats", "grDEvices (>= 3.0.0)"), file = file.path(tmp, "DESCRIPTION"), write = TRUE)
write.dcf(read.dcf(file.path(tmp, "DESCRIPTION")))
# Whoops! we misspelled grDevices. We can replace it.
replace("grDEvices", "grDevices (>= 3.4.0)", file = file.path(tmp, "DESCRIPTION"))

}
